pyproject.toml
[project]
name = "ecom-analytics-pipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "Thu Phan",email = "thuleanhphan@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "pymongo (>=4.16.0,<5.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "ip2location (>=8.11.0,<9.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "h2 (>=4.3.0,<5.0.0)",
    "curl-cffi (>=0.14.0,<0.15.0)",
    "google-cloud-storage (>=3.10.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false